python-analytics/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
textblob==0.17.1
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2
nltk==3.8.1
transformers==4.33.2
torch==2.0.1
google-generativeai==0.1.0
python-dotenv==1.0.0
firebase-admin==6.2.0
plotly==5.15.0
vaderSentiment==3.3.2
spacy==3.6.1
